Revert "media: Update camera stack for 4.19"
This reverts commit d491cdf812c57a2af358363046587141c12cccb0.

+++ drivers/media/platform/msm/camera/cam_icp/fw_inc/hfi_intf.h
@@ -1,6 +1,13 @@
-/* SPDX-License-Identifier: GPL-2.0-only */
-/*
- * Copyright (c) 2017-2019, The Linux Foundation. All rights reserved.
+/* Copyright (c) 2017-2018, The Linux Foundation. All rights reserved.
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License version 2 and
+ * only version 2 as published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
  */
 
 #ifndef _HFI_INTF_H_
@@ -16,10 +23,10 @@
  * @reserved: reserved field
  */
 struct hfi_mem {
-	uint64_t  len;
-	uintptr_t kva;
-	uint32_t  iova;
-	uint32_t  reserved;
+	uint64_t len;
+	uint64_t kva;
+	uint32_t iova;
+	uint32_t reserved;
 };
 
 /**
